requirements.txt
pystow
curies[flask]
pyyaml
